refactor: reorganize type definitions and migrate to _types module
